LDBR-2.30: Добавить проверки на то, имеет ли пользователь доступ к списку карточек и карточке

+++ app/usecases/impl/card.go
@@ -4,10 +4,12 @@ import (
 	"backendServer/app/models"
 	"backendServer/app/repositories"
 	"backendServer/app/usecases"
+	customErrors "backendServer/pkg/errors"
 )
 
 type CardUseCaseImpl struct {
 	cardRepository repositories.CardRepository
+	userRepository repositories.UserRepository
 }
 
 func CreateCardUseCase(cardRepository repositories.CardRepository) usecases.CardUseCase {
@@ -23,20 +25,41 @@ func (cardUseCase *CardUseCaseImpl) CreateCard(card *models.Card) (cid uint, err
 }
 
 func (cardUseCase *CardUseCaseImpl) GetCard(uid, cid uint) (card *models.Card, err error) {
-	// TODO добавить проверку на то, что пользователь имеет доступ к этой карте
+	isAccessed, err := cardUseCase.userRepository.IsCardAccessed(uid, cid)
+	if err != nil {
+		return
+	}
+	if !isAccessed {
+		err = customErrors.ErrNoAccess
+		return
+	}
 
 	card, err = cardUseCase.cardRepository.GetByID(cid)
 	return
 }
 
 func (cardUseCase *CardUseCaseImpl) UpdateCard(uid uint, card *models.Card) (err error) {
-	// TODO добавить проверку на то, что пользователь имеет доступ к этой карте
+	isAccessed, err := cardUseCase.userRepository.IsCardAccessed(uid, card.CID)
+	if err != nil {
+		return
+	}
+	if !isAccessed {
+		err = customErrors.ErrNoAccess
+		return
+	}
 
 	return cardUseCase.cardRepository.Update(card)
 }
 
 func (cardUseCase *CardUseCaseImpl) DeleteCard(uid, cid uint) (err error) {
-	// TODO добавить проверку на то, что пользователь имеет доступ к этой карте
+	isAccessed, err := cardUseCase.userRepository.IsCardAccessed(uid, cid)
+	if err != nil {
+		return
+	}
+	if !isAccessed {
+		err = customErrors.ErrNoAccess
+		return
+	}
 
 	return cardUseCase.cardRepository.Delete(cid)
 }
